refactor: Board 초기화 팩토리 클래스 생성
